Refacotr sobre unqfy, para quitar responsabilidades sobre la playlist
